Move command parser to separate file
