+++ 2/CMakeLists.txt
@@ -4,7 +4,7 @@ project(2)
 set(CMAKE_CXX_STANDARD 17)
 SET(CMAKE_CXX_FLAGS -pthread)
 
-add_executable(2 consumer_producer.h main.cpp function_calculator.h)
+add_executable(2 consumer_producer.h main.cpp function_calculator.h command_parser.h)
 add_executable(tests consumer_producer.h test.cpp)
 
 enable_testing()
